refactor: pageLayout 분리
