--- src/refactoring/components/Layout.tsx
@@ -1,11 +1,0 @@
-interface Props {
-  children: React.ReactNode;
-  title: string;
-}
-
-export const Layout = ({ children, title }: Props) => (
-  <div className="container mx-auto p-4">
-    <h1 className="text-3xl font-bold mb-6">{title}</h1>
-    <div className="grid grid-cols-1 md:grid-cols-2 gap-6">{children}</div>
-  </div>
-);
