+++ src/refactoring/components/PageLayout.tsx
@@ -1,0 +1,15 @@
+import { ReactNode } from 'react';
+
+interface PageLayoutProps {
+  children: ReactNode;
+  title: string;
+}
+
+export const PageLayout = ({ children, title }: PageLayoutProps) => {
+  return (
+    <div className="container mx-auto p-4">
+      <h1 className="text-3xl font-bold mb-6">{title}</h1>
+      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">{children}</div>
+    </div>
+  );
+};
