+++ src/refactoring/components/index.ts
@@ -1,1 +1,2 @@
-export * from './Layout';
+export * from './PageLayout';
+export * from './Navigation';
