+++ src/refactoring/pages/cart/CartPage.tsx
@@ -1,4 +1,5 @@
 import { CartItem, Coupon, Product } from '../../../types.ts';
+import { PageLayout } from '../../components/PageLayout.tsx';
 import { useCart } from '../../hooks';
 
 interface Props {
@@ -41,8 +42,7 @@ export const CartPage = ({ products, coupons }: Props) => {
   };
 
   return (
-    <div className="container mx-auto p-4">
-      <h1 className="text-3xl font-bold mb-6">장바구니</h1>
+    <PageLayout title="장바구니">
       <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
         <div>
           <h2 className="text-2xl font-semibold mb-4">상품 목록</h2>
@@ -190,6 +190,6 @@ export const CartPage = ({ products, coupons }: Props) => {
           </div>
         </div>
       </div>
-    </div>
+    </PageLayout>
   );
 };
